Initial commit malestrom challenge

+++ src/rpc/echo.rs
@@ -30,7 +30,7 @@ pub struct EchoMsgIn {
 }
 
 impl EchoMsgIn {
-    pub fn into_response(&self, outbound_msg_id: usize) -> EchoMsgOut {
+    pub fn into_response(&self, outbound_msg_id: u64) -> EchoMsgOut {
         EchoMsgOut {
             src: self.dest.clone(),
             dest: self.src.clone(),
@@ -38,7 +38,7 @@ impl EchoMsgIn {
         }
     }
 
-    pub fn parse_msg_to_str_response(msg: &str, outbound_msg_id: usize) -> Result<String, errors::ErrorMsg> {
+    pub fn parse_msg_to_str_response(msg: &str, outbound_msg_id: u64) -> Result<String, errors::ErrorMsg> {
         let msg_out = serde_json::from_str::<Self>(msg)
             .map(|m| m.into_response(outbound_msg_id))
             .map_err(|e| errors::ErrorMsg::json_parse_error())?;
@@ -54,17 +54,16 @@ pub struct EchoMsgOut {
 }
 
 
-
 #[derive(Deserialize)]
 pub struct EchoRequestMsg {
     #[serde(rename = "type")]
     typ: Echo,
-    pub msg_id: usize,
+    pub msg_id: u64,
     pub echo: Value
 }
 
 impl EchoRequestMsg {
-    pub fn new(msg_id: usize, echo: Value) -> Self {
+    pub fn new(msg_id: u64, echo: Value) -> Self {
         EchoRequestMsg {
             typ: Echo(MessageType::Echo),
             msg_id,
@@ -72,9 +71,10 @@ impl EchoRequestMsg {
         }
     }
 }
+
 impl rpc::IntoReplyBody for EchoRequestMsg {
     type Item = EchoResponseMsg;
-    fn into_reply(&self, outbound_msg_id: usize) -> EchoResponseMsg {
+    fn into_reply(&self, outbound_msg_id: u64) -> EchoResponseMsg {
         EchoResponseMsg {
             typ: EchoOk(EchoMsgResp::EchoOk),
             msg_id: outbound_msg_id,
@@ -89,9 +89,9 @@ impl rpc::IntoReplyBody for EchoRequestMsg {
 pub struct EchoResponseMsg {
     #[serde(rename = "type")]
     typ: EchoOk,
-    pub msg_id: usize,
+    pub msg_id: u64,
     pub echo: Value,
-    pub in_reply_to: Option<usize>,
+    pub in_reply_to: Option<u64>,
 }
 
 impl rpc::Reply for EchoResponseMsg {}
