+++ src/rpc/gcounter.rs
@@ -1,6 +1,5 @@
 use serde::{Deserialize, Serialize};
 
-
 #[derive(Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 pub enum MsgType {
@@ -14,30 +13,30 @@ pub enum MsgType {
 pub struct AddRequestMsg {
     #[serde(rename = "type")]
     typ: MsgType,
-    msg_id: usize,
-    delta: usize,
+    msg_id: u64,
+    delta: u64,
 }
 
 #[derive(Serialize, Deserialize)]
 pub struct AddResponseMsg {
     #[serde(rename = "type")]
     typ: String,
-    in_reply_to: Option<usize>,
-    msg_id: Option<String>
+    in_reply_to: Option<u64>,
+    msg_id: Option<String>,
 }
 
 #[derive(Serialize, Deserialize)]
 pub struct ReadMsg {
     #[serde(rename = "type")]
     typ: MsgType,
-    msg_id: usize,
+    msg_id: u64,
 }
 
 #[derive(Serialize, Deserialize)]
 pub struct ReadResponseMsg {
     #[serde(rename = "type")]
     typ: String,
-    in_reply_to: Option<usize>,
+    in_reply_to: Option<u64>,
     msg_id: Option<String>,
-    value: usize
+    value: u64,
 }
